+++ src/components/EditionCard/DownloadLink.tsx
@@ -1,7 +1,10 @@
 import { Box, Icon } from "@nypl/design-system-react-components";
+import { useRouter } from "next/router";
 import React from "react";
 import Link from "~/src/components/Link/Link";
+import { LOGIN_LINK_BASE } from "~/src/constants/links";
 import { trackCtaClick } from "~/src/lib/adobe/Analytics";
+import { fulfillFetcher } from "~/src/lib/api/SearchApi";
 import { ItemLink } from "~/src/types/DataModel";
 import { formatUrl } from "~/src/util/Util";
 
@@ -9,12 +12,19 @@ const DownloadLink: React.FC<{
   downloadLink: ItemLink;
   title: string;
   isLoggedIn: boolean;
-}> = ({ downloadLink, title, isLoggedIn }) => {
+  loginCookie?: string;
+}> = ({ downloadLink, title, isLoggedIn, loginCookie }) => {
+  const router = useRouter();
   if (downloadLink && downloadLink.url) {
     let linkText = "Download PDF";
     let linkUrl = formatUrl(downloadLink.url);
 
-    const trackDownloadCta = () => {
+    const handleDownload = async (e) => {
+      if (linkUrl.includes("/fulfill/")) {
+        e.preventDefault();
+        const redirectUrl = await fulfillFetcher(linkUrl, loginCookie);
+        router.push(redirectUrl);
+      }
       trackCtaClick({
         cta_section: `${title}`,
         cta_text: "Download",
@@ -24,9 +34,7 @@ const DownloadLink: React.FC<{
 
     if (downloadLink.flags.nypl_login && !isLoggedIn) {
       linkText = "Log in to download PDF";
-      linkUrl = `https://login.nypl.org/auth/login?redirect_uri=${encodeURIComponent(
-        window.location.href
-      )}`;
+      linkUrl = LOGIN_LINK_BASE + encodeURIComponent(window.location.href);
     }
 
     return (
@@ -34,7 +42,7 @@ const DownloadLink: React.FC<{
         <Link
           to={`${linkUrl}`}
           linkType="buttonSecondary"
-          onClick={trackDownloadCta}
+          onClick={(e) => handleDownload(e)}
           aria-label={`${title} Download PDF`}
         >
           <Icon